+++ fine_tune.py
@@ -12,7 +12,7 @@ from model.vocabulary import *
 from model.model import *
 from model.dataset import *
 from model import utils
-from randomize_smiles import randomize_smiles
+from model.randomize_smiles import randomize_smiles
 
 rdBase.DisableLog("rdApp.error")
